pyproject.toml
[tool.poetry]
name = "tool"
version = "0.1.0"
description = ""
authors = ["Pooria Azami <pooriaazami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
